fix: replace 208-line redundant test-runner ordering tests with minimal afterEach cleanup test (#364)
Closes #364
